@@ -14,7 +14,7 @@ description: A new Flutter project.
 version: 1.2.1+220408
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.2.0 <3.0.0"
 
 dependency_overrides:
   photo_view:
